requirements.txt
tkinter
keyboard>=0.13.5
pystray>=0.19.5
pillow>=11.2.1
pyinstaller>=6.13.0